FEATURE/UpdateMenuItemUseCase (#22)
* [ FEAT ] UpdateMenuitemUseCase

* [ FIX ] UpdateMenuItem
- Change in the MenuItem entity to receive a null price and change it to decimal

+++ Application/UseCases/MenuItems/v1/GetMenuItemUseCase/Models/MenuItemResponse.cs
@@ -7,7 +7,7 @@ namespace Application.UseCases.MenuItems.v1.GetMenuItemUseCase.Models
         public Guid Id { get; set; } = menuItem.Id;
         public string Name { get; set; } = menuItem.Name;
         public string Description { get; set; } = menuItem.Description;
-        public double Price { get; set; } = menuItem.Price;
+        public decimal Price { get; set; } = menuItem.Price;
 
         public IEnumerable<MenuItemImagesResponse> Images { get; set; } = menuItem.Images.Select(item => new MenuItemImagesResponse(item));
     }
